Refactoring.

+++ inc/cpu/usart.h
@@ -20,17 +20,17 @@ typedef struct USART_Definition
 	uint32_t GpioPortClock;
 	GPIO_TypeDef *GpioPort;
 
-	uint16_t RxPin;
-	uint16_t TxPin;
+	uint32_t RxPin;
+	uint32_t TxPin;
 
 	uint8_t RxPinSource;
 	uint8_t TxPinSource;
 } USART_Definition;
 
-bool USART_Enable(USART_TypeDef *usartx, uint32_t baudRate);
+bool USART_Enable(USART_TypeDef *usartx, unsigned int baudRate);
 bool USART_Disable(USART_TypeDef *usartx);
 void USART_SendChar(USART_TypeDef *usartx, char c);
-uint32_t USART_SendString(USART_TypeDef *usartx, const char *str);
+int USART_SendString(USART_TypeDef *usartx, const char *str);
 USART_Definition *USART_GetDefinition(USART_TypeDef *usartx);
 
 #ifdef __cplusplus
